+++ spec/webdriver_firefox_spec.rb
@@ -15,8 +15,9 @@ describe 'Webdriver Firefox' do
   end
 
   it 'Webdriver#click_on_locator_coordinates correct for .0 values' do
-    webdriver.open('www.google.com')
-    webdriver.click_on_locator_coordinates('//*[@id="body"]', 20.0, 10)
+    webdriver.open("file://#{Dir.pwd}/spec/html_examples/"\
+                      'web_driver_element_size_by_js_spec.html')
+    webdriver.click_on_locator_coordinates('//div', 20.0, 10)
   end
 
   it 'Check That firefox return empty browser logs' do