+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ### Changes
 
 * Update `geckodriver` to `0.28.0`
+* Fix firefox `Webdriver#click_on_locator_coordinates` test
+  to not rely on google
 
 ## 0.7.0 (2020-11-18)